feat(describeTable): support autoIncrement for mysql
