+++ lib/dialects/postgres/query-generator.js
@@ -100,29 +100,27 @@ class PostgresQueryGenerator extends AbstractQueryGenerator {
   }
 
   describeTableQuery(tableName, schema) {
-    if (!schema) {
-      schema = 'public';
-    }
-    return 'SELECT pk.constraint_type as "Constraint", c.column_name as "Field", ' +
-              'c.column_default as "Default", c.is_nullable as "Null", ' +
-              '(CASE WHEN c.udt_name = \'hstore\' THEN c.udt_name ELSE c.data_type END) || (CASE WHEN c.character_maximum_length IS NOT NULL THEN \'(\' || c.character_maximum_length || \')\' ELSE \'\' END) as "Type", ' +
-              '(SELECT array_agg(e.enumlabel) ' +
-              'FROM pg_catalog.pg_type t JOIN pg_catalog.pg_enum e ON t.oid=e.enumtypid ' +
-              'WHERE t.typname=c.udt_name) AS "special", ' +
-              '(SELECT pgd.description FROM pg_catalog.pg_statio_all_tables AS st ' +
-              'INNER JOIN pg_catalog.pg_description pgd on (pgd.objoid=st.relid) ' +
-              'WHERE c.ordinal_position=pgd.objsubid AND c.table_name=st.relname) AS "Comment" ' +
-            'FROM information_schema.columns c ' +
-            'LEFT JOIN (SELECT tc.table_schema, tc.table_name, ' +
-              'cu.column_name, tc.constraint_type ' +
-              'FROM information_schema.TABLE_CONSTRAINTS tc ' +
-              'JOIN information_schema.KEY_COLUMN_USAGE  cu ' +
-              'ON tc.table_schema=cu.table_schema and tc.table_name=cu.table_name ' +
-                'and tc.constraint_name=cu.constraint_name ' +
-                'and tc.constraint_type=\'PRIMARY KEY\') pk ' +
-            'ON pk.table_schema=c.table_schema ' +
-            'AND pk.table_name=c.table_name ' +
-            'AND pk.column_name=c.column_name ' +
+    if (!schema) schema = 'public';
+    
+    return 'SELECT ' +
+      'pk.constraint_type as "Constraint",' +
+      'c.column_name as "Field", ' +
+      'c.column_default as "Default",' +
+      'c.is_nullable as "Null", ' +
+      '(CASE WHEN c.udt_name = \'hstore\' THEN c.udt_name ELSE c.data_type END) || (CASE WHEN c.character_maximum_length IS NOT NULL THEN \'(\' || c.character_maximum_length || \')\' ELSE \'\' END) as "Type", ' +
+      '(SELECT array_agg(e.enumlabel) FROM pg_catalog.pg_type t JOIN pg_catalog.pg_enum e ON t.oid=e.enumtypid WHERE t.typname=c.udt_name) AS "special", ' +
+      '(SELECT pgd.description FROM pg_catalog.pg_statio_all_tables AS st INNER JOIN pg_catalog.pg_description pgd on (pgd.objoid=st.relid) WHERE c.ordinal_position=pgd.objsubid AND c.table_name=st.relname) AS "Comment" ' +
+      'FROM information_schema.columns c ' +
+      'LEFT JOIN (SELECT tc.table_schema, tc.table_name, ' +
+      'cu.column_name, tc.constraint_type ' +
+      'FROM information_schema.TABLE_CONSTRAINTS tc ' +
+      'JOIN information_schema.KEY_COLUMN_USAGE  cu ' +
+      'ON tc.table_schema=cu.table_schema and tc.table_name=cu.table_name ' +
+      'and tc.constraint_name=cu.constraint_name ' +
+      'and tc.constraint_type=\'PRIMARY KEY\') pk ' +
+      'ON pk.table_schema=c.table_schema ' +
+      'AND pk.table_name=c.table_name ' +
+      'AND pk.column_name=c.column_name ' +
       `WHERE c.table_name = ${this.escape(tableName)} AND c.table_schema = ${this.escape(schema)} `;
   }
 
@@ -856,22 +854,22 @@ class PostgresQueryGenerator extends AbstractQueryGenerator {
    */
   _getForeignKeyReferencesQueryPrefix() {
     return 'SELECT ' +
-        'DISTINCT tc.constraint_name as constraint_name, ' +
-        'tc.constraint_schema as constraint_schema, ' +
-        'tc.constraint_catalog as constraint_catalog, ' +
-        'tc.table_name as table_name,' +
-        'tc.table_schema as table_schema,' +
-        'tc.table_catalog as table_catalog,' +
-        'kcu.column_name as column_name,' +
-        'ccu.table_schema  AS referenced_table_schema,' +
-        'ccu.table_catalog  AS referenced_table_catalog,' +
-        'ccu.table_name  AS referenced_table_name,' +
-        'ccu.column_name AS referenced_column_name ' +
+      'DISTINCT tc.constraint_name as constraint_name, ' +
+      'tc.constraint_schema as constraint_schema, ' +
+      'tc.constraint_catalog as constraint_catalog, ' +
+      'tc.table_name as table_name,' +
+      'tc.table_schema as table_schema,' +
+      'tc.table_catalog as table_catalog,' +
+      'kcu.column_name as column_name,' +
+      'ccu.table_schema  AS referenced_table_schema,' +
+      'ccu.table_catalog  AS referenced_table_catalog,' +
+      'ccu.table_name  AS referenced_table_name,' +
+      'ccu.column_name AS referenced_column_name ' +
       'FROM information_schema.table_constraints AS tc ' +
-        'JOIN information_schema.key_column_usage AS kcu ' +
-          'ON tc.constraint_name = kcu.constraint_name ' +
-        'JOIN information_schema.constraint_column_usage AS ccu ' +
-          'ON ccu.constraint_name = tc.constraint_name ';
+      'JOIN information_schema.key_column_usage AS kcu ' +
+      'ON tc.constraint_name = kcu.constraint_name ' +
+      'JOIN information_schema.constraint_column_usage AS ccu ' +
+      'ON ccu.constraint_name = tc.constraint_name ';
   }
 
   /**
